Stay

+++ src/routes/stay/providers/stay.service.ts
@@ -11,7 +11,14 @@ import { ErrorMsg } from "../../../common/mapper/error";
 import { Gender, Grade, UserJWT } from "../../../common/mapper/types";
 import { safeFindOne } from "../../../common/utils/safeFindOne.util";
 import { isInRange } from "../../../common/utils/staySeat.util";
-import { Stay, StayApply, StayOuting, StaySeatPreset, User } from "../../../schemas";
+import {
+  Stay,
+  StayApply,
+  StayOuting,
+  StaySeatPreset,
+  User,
+  StayApplyPeriod_Stay,
+} from "../../../schemas";
 import { UserManageService } from "../../user/providers";
 import {
   AddStayOutingDTO,
@@ -43,12 +50,7 @@ export class StayService {
 
     const stays = await this.stayRepository
       .createQueryBuilder("stay")
-      .innerJoin(
-        "stay.stay_apply_period",
-        "stay_apply_period",
-        "stay_apply_period.grade = :grade AND stay_apply_period.apply_start <= :now AND stay_apply_period.apply_end >= :now",
-        { grade: data.grade, now },
-      )
+      .innerJoin("stay.stay_apply_period", "stay_apply_period")
       .leftJoin("stay.stay_apply", "stay_apply")
       .leftJoin("stay_apply.user", "user")
       .leftJoin("stay.stay_seat_preset", "stay_seat_preset")
@@ -67,6 +69,7 @@ export class StayService {
         "user.id",
         "user.name",
       ])
+      .orderBy("stay.stay_from", "ASC")
       .getMany();
 
     return stays.map((stay) => ({
@@ -94,16 +97,21 @@ export class StayService {
 
   async createStayApply(user: UserJWT, data: CreateUserStayApplyDTO) {
     const target = await safeFindOne<User>(this.userRepository, user.id);
-    const stay = await safeFindOne<Stay>(this.stayRepository, {
-      where: {
-        id: data.stay,
-        stay_apply_period: {
-          grade: data.grade,
-          apply_start: LessThanOrEqual(new Date()),
-          apply_end: MoreThanOrEqual(new Date()),
+
+    const stay = await safeFindOne<Stay>(
+      this.stayRepository,
+      {
+        where: {
+          id: data.stay,
+          stay_apply_period: {
+            grade: data.grade,
+            apply_start: LessThanOrEqual(new Date()),
+            apply_end: MoreThanOrEqual(new Date()),
+          },
         },
       },
-    });
+      new HttpException(ErrorMsg.Stay_NotInApplyPeriod(), HttpStatus.NOT_FOUND),
+    );
 
     const exists = await this.stayApplyRepository.findOne({
       where: { user: target, stay: { id: data.stay } },
@@ -171,6 +179,10 @@ export class StayService {
     if (stayApply.user.id !== user.id)
       throw new HttpException(ErrorMsg.PermissionDenied_Resource(), HttpStatus.FORBIDDEN);
 
+    if (!(await this.validateStayPeriod(user, data.grade, stayApply.stay.stay_apply_period))) {
+      throw new HttpException(ErrorMsg.Stay_NotInApplyPeriod(), HttpStatus.FORBIDDEN);
+    }
+
     const staySeatCheck = await this.stayApplyRepository.findOne({
       where: { stay_seat: data.stay_seat.toUpperCase(), stay: { id: stayApply.stay.id } },
     });
@@ -225,9 +237,23 @@ export class StayService {
   }
 
   async deleteStayApply(user: UserJWT, data: StayIdDTO) {
-    const stayApply = await safeFindOne<StayApply>(this.stayApplyRepository, data.id);
+    const now = new Date();
+
+    const stayApply = await this.stayApplyRepository.findOne({
+      where: { id: data.id },
+      relations: ["stay", "stay.stay_apply_period"],
+    });
+    if (!stayApply) throw new HttpException(ErrorMsg.Resource_NotFound(), HttpStatus.NOT_FOUND);
+
     if (stayApply.user.id !== user.id)
       throw new HttpException(ErrorMsg.PermissionDenied_Resource(), HttpStatus.FORBIDDEN);
+
+    if (!(await this.userManageService.checkUserDetail(user.email, { grade: data.grade })))
+      throw new HttpException(ErrorMsg.PermissionDenied_Resource(), HttpStatus.FORBIDDEN);
+
+    if (!(await this.validateStayPeriod(user, data.grade, stayApply.stay.stay_apply_period))) {
+      throw new HttpException(ErrorMsg.Stay_NotInApplyPeriod(), HttpStatus.FORBIDDEN);
+    }
 
     return this.stayApplyRepository.remove(stayApply);
   }
@@ -248,6 +274,11 @@ export class StayService {
       },
       relations: { stay: true },
     });
+
+    // verification period
+    if (!(await this.validateStayPeriod(user, data.grade, apply.stay.stay_apply_period))) {
+      throw new HttpException(ErrorMsg.Stay_NotInApplyPeriod(), HttpStatus.FORBIDDEN);
+    }
 
     console.log(apply.user.id);
     if (apply.user.id !== target.id)
@@ -274,6 +305,10 @@ export class StayService {
         )) ||
       null;
 
+    if (outing.from >= outing.to) {
+      throw new HttpException(ErrorMsg.ProvidedTime_Invalid(), HttpStatus.BAD_REQUEST);
+    }
+
     const saved = await this.stayOutingRepository.save(outing);
     return await safeFindOne<StayOuting>(this.stayOutingRepository, saved.id);
   }
@@ -282,13 +317,20 @@ export class StayService {
     const target = await safeFindOne<User>(this.userRepository, user.id);
     const outing = await safeFindOne<StayOuting>(this.stayOutingRepository, {
       where: { id: data.outing_id },
-      relations: { stay_apply: { user: true, stay: true } },
+      relations: { stay_apply: { user: true, stay: { stay_apply_period: true } } },
       loadEagerRelations: false,
     });
 
     console.log(outing.stay_apply.user.id);
     if (outing.stay_apply.user.id !== target.id)
       throw new HttpException(ErrorMsg.PermissionDenied_Resource(), HttpStatus.FORBIDDEN);
+
+    // verification period
+    if (
+      !(await this.validateStayPeriod(user, data.grade, outing.stay_apply.stay.stay_apply_period))
+    ) {
+      throw new HttpException(ErrorMsg.Stay_NotInApplyPeriod(), HttpStatus.FORBIDDEN);
+    }
 
     outing.reason = data.outing.reason;
     outing.breakfast_cancel = data.outing.breakfast_cancel;
@@ -309,6 +351,10 @@ export class StayService {
         )) ||
       null;
 
+    if (outing.from >= outing.to) {
+      throw new HttpException(ErrorMsg.ProvidedTime_Invalid(), HttpStatus.BAD_REQUEST);
+    }
+
     const saved = await this.stayOutingRepository.save(outing);
     return await safeFindOne<StayOuting>(this.stayOutingRepository, saved.id);
   }
@@ -317,7 +363,7 @@ export class StayService {
     const target = await safeFindOne<User>(this.userRepository, user.id);
     const outing = await safeFindOne<StayOuting>(this.stayOutingRepository, {
       where: { id: data.id },
-      relations: { stay_apply: { user: true } },
+      relations: { stay_apply: { user: true, stay: { stay_apply_period: true } } },
       loadEagerRelations: false,
     });
 
@@ -325,6 +371,11 @@ export class StayService {
     if (outing.stay_apply.user.id !== target.id)
       throw new HttpException(ErrorMsg.PermissionDenied_Resource(), HttpStatus.FORBIDDEN);
 
+    if (
+      !(await this.validateStayPeriod(user, data.grade, outing.stay_apply.stay.stay_apply_period))
+    ) {
+      throw new HttpException(ErrorMsg.Stay_NotInApplyPeriod(), HttpStatus.FORBIDDEN);
+    }
     return await this.stayOutingRepository.remove(outing);
   }
 
@@ -345,5 +396,41 @@ export class StayService {
             !preset.only_readingRoom,
         ) && (await this.userManageService.checkUserDetail(user.email, { gender, grade }))
     );
+  }
+
+  private async validateStayPeriod(
+    user: UserJWT,
+    grade: Grade,
+    stay_apply_period: StayApplyPeriod_Stay[],
+  ) {
+    let isSame = true;
+    let last = "";
+    for (const period of stay_apply_period) {
+      if (last === "") {
+        last = period.apply_start.getTime().toString() + period.apply_end.getTime().toString();
+        continue;
+      }
+      if (
+        last !==
+        period.apply_start.getTime().toString() + period.apply_end.getTime().toString()
+      ) {
+        isSame = false;
+        break;
+      }
+      last = period.apply_start.getTime().toString() + period.apply_end.getTime().toString();
+    }
+
+    if (!isSame) {
+      const success = await this.userManageService.checkUserDetail(user.email, { grade: grade });
+      if (!success)
+        throw new HttpException(ErrorMsg.PermissionDenied_Resource_Grade(), HttpStatus.FORBIDDEN);
+    }
+
+    const now = new Date();
+    const validPeriod = stay_apply_period.find(
+      (p) => p.grade === Number(grade) && p.apply_start <= now && p.apply_end >= now,
+    );
+
+    return !!validPeriod;
   }
 }
